Hack to get SMBIOS from kernel log.
This implements a hack to get the SMBIOS base address from the kernel log. It
is useful in some situations where the SMBIOS entry point structure does not
appear in the legacy region, which is sometimes the case with EFI BIOSes.

TODO: Is there a "proper" EFI variable fix for this?



git-svn-id: https://mosys.googlecode.com/svn/trunk@55 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e