Switch the AGZ's probing method to use an experimental HWID doohicky in ACPI.
The HWID is basically a GUID that is assigned arbitrarily to the platform. It's
basically like the SMBIOS system info strings, hashed into something completely
unreadable and otherwise worthless. We'll use it for obfuscation for now and
ditch it once the juicy system details are made available in SMBIOS type 1
tables.



git-svn-id: https://mosys.googlecode.com/svn/trunk@107 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e