Add initial support for variable-length blobs.
The idea here is that if the blob has a length of zero specified in the table,
but has a valid pointer and all, we will assume the length is from the blob
offset to the rest of the ROM size. 

TODO: This is kind of funky, and the "size" field gets printed incorrectly.



git-svn-id: https://mosys.googlecode.com/svn/trunk@214 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e