Another attempt to mitigate alex ec passthru-mode related failures.
This adds a call to alex_ec_exit_passthru_mode() as part of the EC
setup sequence. The theory is that exiting passthru mode first might be
more reliable than attempting to enter passthru mode, failing, and then
attempting to exit. If we are not in passthru mode when the first
call to exit_passthru_mode is called, it is benign.



git-svn-id: https://mosys.googlecode.com/svn/trunk@282 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e