Make all x86 platforms return canonical name instead of SMBIOS-provided name.
SMBIOS was used as the standard earlier on since it was assumed to be available
and stable, and before we had alternative methods of probing platforms.

Since we can use HWID as an identifier, it's easier to keep the platform name
stable even if we change BIOS. This is especially useful for early bring-up,
e.g. using SMBIOS may be incorrect or unavailable.



git-svn-id: https://mosys.googlecode.com/svn/trunk@380 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e