Fix sensor command so it doesn't erroneously print all sensors.
This fixes the commands to print various sensor types. The sensor_print_exec()
function was receiving bogus arguments earlier, and it just so happened that
the default behavior was to print all sensors. This resulted in all sensors
being displayed regardless of whether the user specified fantach, thermal, or
voltage.

The new behavior is to return -ENOSYS, exit with non-zero, and print "Command
not supported on this platform" if the user attempts to print a sensor type
which is not supported.



git-svn-id: https://mosys.googlecode.com/svn/trunk@408 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e