Revamp the sensor model to pass around a new sensor_reading struct
This adds a new struct which can be used to store a sensor reading in addition
to some other useful information, such as the mode which the sensor is in.  It
might also be used later on to store information such as thresholds.

In the long run, it might be useful to add sensor value types such as int,
double, etc. But for now, it's still assumed that the value is a double.



git-svn-id: https://mosys.googlecode.com/svn/trunk@410 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e