Add support for retrieving embedded SPD info for Lumpy
Lumpy has an on-board memory module as well as a slotted memory
module. The on-board module does not have a physical SPD; instead
the SPD data is stored on the BIOS ROM as a CBFS entry. This patch
is a first-attempt at retreiving that info.

TODO: Add support to select which SPD entry to choose from in CBFS.



git-svn-id: https://mosys.googlecode.com/svn/trunk@470 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e