ELOG: Add support for printing multiple entries from ME
If the management engine is disabled or reports an error then
a series of 7 bytes are logged into a single event.  This can
then be parsed to provide more detail on why the ME reported
a problem.

To handle this the single event is split into multiple entries
in the log output.

BUG=chrome-os-partner:12766
BRANCH=link,parrot,butterfly,stout
TEST=manual

1) Disable the ME with override pin on Parrot device
2) Boot the system and dump the event log

49 | 2012-09-06 14:29:12 | Management Engine | Disabled
50 | 2012-09-06 14:29:12 | ME Working State | Normal
51 | 2012-09-06 14:29:12 | ME Operation State | Bring up
52 | 2012-09-06 14:29:12 | ME Operation Mode | Security Override via Jumper
53 | 2012-09-06 14:29:12 | ME Progress Phase | BUP Phase
54 | 2012-09-06 14:29:12 | ME PM Event | Clean Moff->Mx wake

Change-Id: I915ac4db2cbb2abf3fbfbe2337c6b3d91b19a6d2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32443
Reviewed-by: David Hendricks <dhendrix@chromium.org>