add a return code to kv_pair_print()
This adds a return code to kv_pair_print(). This is in anticipation
of adding a single-value printing mode which can return a failure
code if the key specified by the user is not matched.

Call sites are also updated to take return code into account. This
meant updating some functions to return an int instead of void.

(this also fixes a trivial >80 cols error)

BRANCH=none
BUG=chromium-os:36425
TEST=tested in follow-up CL

Change-Id: Ic5a22c1db018c576d383d39938c76bd2163b6228
Reviewed-on: https://gerrit.chromium.org/gerrit/39749
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>