Re-factor kv_pair_print_to_file() to reduce nesting
kv_pair_print_to_file() iterates over all kv_pairs in the provided
list for all styles, so each case has its own for-loop and
if-statement for sanity checking that a key and value exist.

This flips that around so that there is a single for-loop and
if-statement, reducing duplicate statements and reducing nesting by
1 level.

BRANCH=none
BUG=none
TEST=ran "mosys ec info" on stout with all formatting flags, saw
no change in output format.

Change-Id: I24724e8f6135042ed8ee4deb026bb50dff849316
Reviewed-on: https://gerrit.chromium.org/gerrit/39750
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>