fix a trailing newline at end of output in -l mode
The long style formatting prints a newline unconditionally at the end
of each line. Since a newline is printed at the end of
kv_pair_print_to_file(), two newlines appear at the end of the output.

This makes long style formatting print a newline only if it's not
at the end of the list. The final newline is inserted at the end
of the function, like with other styles.

BRANCH=none
BUG=none
TEST=ran "mosys -l platform name" and no longer saw a newline at the
end of the output

Change-Id: Ia7781aed871754be04e14e63da432b3994d118c6
Reviewed-on: https://gerrit.chromium.org/gerrit/39751
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>