use a more sensible limit for line length when printing tree info
This defines a maximum line length to use when recursively adding
command info which we can then use to calculate the maximum number
of bytes remaining on each iteration.

The line length must allow for leading tabs, the type of command,
the entire command, and a string terminator. The longest command
to date uses 43 characters in verbose mode.

BRANCH=none
BUG=none
TEST=run on stout

Change-Id: I0fd2210369301e5555a2c91678c15f50142a3547
Reviewed-on: https://gerrit.chromium.org/gerrit/41311
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>