elog: Explicitly include mosys/platform.h in intf/mmio.h.
The inline functions in intf/mmio.h actually dereference the intf parameter
they take, so there has to be an actual definition for that type, not just a
declaration that it exists.

BUG=chrome-os-partner:16132
TEST=With this and other changes, ran mosys eventlog list and mosys smbios
info log on spring. On link, ran mosys eventlog list. Built for parrot, stout,
and butterfly.
BRANCH=spring

Change-Id: Ibccc18a1035f25eeb893f389fab08c8d33278743
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/49865
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>