eeprom: add read_by_name() device method
This adds a method to the eeprom device struct which will allow the
caller to specify a named region (in an fmap, for example) to read.
The implementation will find the region and allocate + fill memory.

BUG=chrome-os-partner:16132
BRANCH=spring
TEST=ran "mosys eventlog list" on Spring with two firmware images,
one with actual eventlog and one without.

Change-Id: I58937cd7fcf173dcd563ae69a352980506d5b042
Reviewed-on: https://gerrit.chromium.org/gerrit/50425
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>