elog: Add new events
Coreboot was recently extended to keep track of some extra
device info in CMOS similar to how it handles post codes.
Currently this consists of the device path of the device
that coreboot is about to initialize.

This is encapsualted in a new event type that mosys can
decode.

Also, the "EC Shutdown" event type was added in a previous
platform but was never added to mosys.

BUG=chrome-os-partner:19980
BRANCH=none
TEST=manual: interrupted boot with reset button and
gathered the eventlog.  Mosys has been extended to
decode the well-known POST codes:

26 | 2013-06-10 10:32:48 | System boot | 120
27 | 2013-06-10 10:32:48 | Last post code in previous boot | 0x75 | Device Initialize
28 | 2013-06-10 10:32:48 | Extra info from previous boot | PCI | 00:16.0
29 | 2013-06-10 10:32:48 | Reset Button
30 | 2013-06-10 10:32:48 | System Reset

Change-Id: I08f4971b2ced78b56b560b4f2174a99f854000bf
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58252
Reviewed-by: David Hendricks <dhendrix@chromium.org>