add "-s" and "-S" back into help menu
Commit bded98 seems to have clobbered some previous changes that had
been made to the help menu.

Since this is a formatting flag, I also moved it up next to the
descriptions for -k and -l so hppefully it's a bit easier for users
to find.

BUG=none
BRANCH=none
TEST=run mosys -h, see that -s and -S both show up

Change-Id: Iaefb82c351693c792f654d4e8ccadfb5a7c047b9
Reviewed-on: https://gerrit.chromium.org/gerrit/60543
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>