eventlog: Change the signature of the "add" callback and subcommand.
The eventlog subcommand took a single, nebulus "event" parameter, and the
callback took the values argc and argv with no interpretation. Now, the
subcommand takes one or three parameters, first the type of the new event,
and then optionally the size and contents of the data parameter. The type and
size are both text representations of numeric values, and the data is treated
as a binary blob and pasted into the event directly.

BUG=chrome-os-partner:16132
TEST=Built for spring.
BRANCH=spring

Change-Id: I75b9fb4b6b3177273373df25a1293a337c738ab9
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62722
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>