spring: Enable the "add" subcommand by installing a callback.
The callback is the function added in the previous CL.

BUG=chrome-os-partner:16132
TEST=Ran "mosys eventlog add" and then "mosys eventlog list" to and verified
that an event was added. Ran "mosys eventlog add" and "mosys eventlog list" in
a loop and verified that the events were added successfully, and that when the
log filled up enough that it was partially cleared successfully and that a
clear event was added.
BRANCH=spring

Change-Id: Iba3b26c8d3b0e24ab1bb64c65fa9ec2d24024709
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62724
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>