Add a function to probe using device-tree compatible node
The device-tree model node is no longer the preferred method of
probing on systems which use device tree since it may be too specific
to a particular board.

The new way to ID a platform using device tree is to use the
"compatible" node which contains multiple (possibly several) strings
with identifiers of various specificity.

BUG=chrome-os-partner:22683
BRANCH=none
TEST=Tested on Snow with new kernel and follow-up CL

Change-Id: Ia1bf679cfdc2bc0fa79b5adc6a107b8927c70594
Reviewed-on: https://chromium-review.googlesource.com/169376
Commit-Queue: David Hendrix <dhendrix@chromium.org>
Tested-by: David Hendrix <dhendrix@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>