exynos5420: Fill out table of GPIO banks
This adds missing entries to the table of GPIO banks so that
"mosys gpio list" prints all GPIOs.

This isn't used for anything in production, but is handy for
validating that GPIO states are set and act in a sane manner
during bring-up.

BUG=chrome-os-partner:22524
BRANCH=none
TEST=ran "mosys -k gpio list" on kirby

Change-Id: I3b2ae8711097e2beae4d218b4ad7919512098a20
Reviewed-on: https://chromium-review.googlesource.com/170232
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: David Hendrix <dhendrix@chromium.org>
Commit-Queue: David Hendrix <dhendrix@chromium.org>