Remove default platform support.
This removes "default platform" support and makes mosys exit if
the platform cannot be detected. The default platform was
originally intended to allow mosys to provide some generic
functionality, but it was never implemented for ARM and seemed
somewhat detrimental to development for x86.

We may revisit this in the future, in which case we'll need to
provide default platform support for both ARM and x86.

BUG=chrome-os-partner:24492
BRANCH=none
TEST=Tested on pit. Setting CONFIG_PLATFORM_PEACH=y made things
work as normal. Unsetting CONFIG_PLATFORM_PEACH caused mosys to
report "Platform not supported" and exit gracefully. Overriding
with "-p SMDK5250" worked, but of course that is something that
should not be done under normal circumstances.

Change-Id: I5ea9856701be6971fc79a73badd0e3ae26de4fce
Reviewed-on: https://chromium-review.googlesource.com/180029
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>