it8772: Add option to force tach read of fan RPM
If available, the tach input should be more accurate than the PWM output
to measure fan speed. Add option to force reading of the tach,
regardless of whether we are in auto mode or not.

BUG=chrome-os-partner:24363
TEST=Manual on Monroe. Enable force_read_tach, verify that RPM is
reported correctly for high and near-zero RPM.
BRANCH=None.

Change-Id: I3ad4e16b79f4c8cf7bec33c467af72237172b0a4
Reviewed-on: https://chromium-review.googlesource.com/180034
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>