Allow for partial matching of FDT compat strings
This adds an argument to probe_fdt_compatible() so that the caller
can specify whether or not to allow partial matches of the supplied
platform ID and the FDT compat string. This is useful if we expect
minor revisions that don't require more explicit probing.

BUG=none
BRANCH=none
TEST=tested with follow-up Storm patch

Change-Id: I7f42cc424514433e465a52bc0e6c08dbb0b0d371
Reviewed-on: https://chromium-review.googlesource.com/215390
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>