lib: Add memory info functions that don't rely on SPD
This introduces some new library functions to help reduce pain of
supporting memory info commands on systems that don't have SPDs.

For a while I was hoping that memory vendors would be polite enough
to provide SPD data for their chips, even when the chips themselves
don't have SPDs. This is common in the x86 world. But for ARM systems
we've had to generate the SPD-style data ourselves which is rather
tedious.

This attempts to allow us to provide data in more free-form manner.

BUG=chrome-os-partner:32803
BRANCH=none
TEST=tested with follow-up patches

Change-Id: I8d8f95c0725beccfce8a918ce0e12ba2610862b4
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223031
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>