Add "legacy_ec_cb" interface for older platforms
The EC callbacks are getting re-factored to make it easier to support
devices with multiple ECs.

Re-factoring some of the old interfaces involves a bit more work than
just changing function args. So this renames ec_cb to legacy_ec_cb
and updates old platforms to use it and avoid disruption.

BUG=chrome-os-partner:32962
BRANCH=None
TEST=tested "mosys ec info" on stout

Change-Id: I3916ce350bde2bceac92863c0e650ada93e1bb3f
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/228953
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>