Check if intf->cb->ec->destroy exists before calling it
Some systems use an EC destroy callback, for example, to close device
nodes in /dev. Currently this is the case for CrOS ECs which we
access via devfs.

However if platform setup is unsuccessful, for example if the CrOS
EC device node doesn't exist, then the destroy callback will still
be called and cause mosys to segfault.

BUG=chromium:437951
BRANCH=none
TEST=renamed /dev/cros_ec on nyan, ran mosys and it didn't segfault

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: Iaff9553352d1a011390d4bd47c42ea792c6d64f8
Reviewed-on: https://chromium-review.googlesource.com/234042
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>