elog: Check for zero-length entries in parsing logic
This adds checks to log entry iterating logic used in mosys eventlog
add, list, and clear commands to avoid problems in case a zero-length
entry is encountered. If one is encountered when adding or
listing events, we simply quit with an error message. For for clearing
commands, we print an error message and proceed.

BUG=chrome-os-partner:36337
BRANCH=none
TEST=see below
1. Add an event such that the length rolls over:
   mosys eventlog add 0 `yes '1'|head -494|tr -d '\n'`
2. "mosys eventlog list" errors out with a warning
3. "mosys eventlog add" errors out and prints a warning
4. "mosys eventlog clear" prints warning and clears the log
5. After clearing the log, adding and listing events works again.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I0380cd53651aa82dece5d5bae2b6e72bc24b0d13
Reviewed-on: https://chromium-review.googlesource.com/248260
Reviewed-by: Shawn N <shawnn@chromium.org>