eventlog: Only check regions for eventlog flag
In an earlier implementation of the eventlog code, we checked the
firmware eeprom itself to see if an eventlog existed on it. Later on
we updated the eeprom usage model to name specific regions we're
interested in along with their respective flags, such that we have
flexibility in case a region's name ever changes.

This updates elog_find_log_in_flash() to look at individual regions,
and removes redundant EEPROM_FLAG_EVENTLOG assignments in platform
code.

BUG=none
BRANCH=none
TEST="mosys eventlog list" still works on storm and veyron_speedy

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: If89392c0ec240ff14b9c362bf56d63bf8561e3b6
Reviewed-on: https://chromium-review.googlesource.com/265260
Reviewed-by: Shawn N <shawnn@chromium.org>