clang: Fix an "autological" compare
The argument for "mosys sensor set fantach" is a percentage. The
validity check compares if it's >100 or <0, but since we use an
unsigned type clang spits out an "autological compare" warning.

This updates the check to utilize the second (endptr) argument for
strtoul() and only check if the value is above 100.

BUG=chromium:475959
BRANCH=none
TEST=Compiled for peppy, no longer see the warning.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I3e8d261d18eaea7abf21047e7bc6fa2caf4a611b
Reviewed-on: https://chromium-review.googlesource.com/265330
Reviewed-by: Shawn N <shawnn@chromium.org>