spd: Remove dead code from SPD library
This removes a lot of dead code for printing detailed SPD info. I'll
be sad to see it go, but I can't imagine it being particularly useful
at this point and we haven't bothered to do this for DDR3/4. There
are also some bugs that clang picked up on.

Some of the memory controllers that are used in CrOS platforms support
DDR2, so we can leave support for decoding some specific fields. A
follow-up patch will take care of remaining bugs found by clang.

BUG=chromium:475959
BRANCH=none
TEST=compiled for peppy (with gcc) and saw a lot of warnings go away
with clang.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I3d6fd4eaa5d17f97643021d7e54f3e7a075846d9
Reviewed-on: https://chromium-review.googlesource.com/265339
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>