pinky: Enable EC callbacks by default, disable if no EC
Commit 799f42 made it so that intf->cb->ec would be set up only if
an EC was detected. Catch-22: intf->cb->ec is required for probing
if the platform does have an EC. D'oh!

This inverts the earlier logic to set the intf->cb->ec by default
and set it to NULL only if the platform is known to not have an EC.

BUG=chrome-os-partner:39280
BRANCH=veyron
TEST=tested on _both_ brain and speedy this time...

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: Iacba988b637d8aeb41cfb7d35abc82b026bdda6c
Reviewed-on: https://chromium-review.googlesource.com/266207
Reviewed-by: Shawn N <shawnn@chromium.org>