elog: check that region entry exists when iterating thru ROMs
This adds a check in elog_find_log_in_flash() to ensure that the
"region" member of the eeprom struct is non-NULL before attempting
to check its flags.

We never encountered this issue in the past because platforms with
ROMs listed put the host firmware ROM first which is where the
eventlog region would typically be found. CL:294332 changed this
convention for Veyron platforms for those with an EC.

BUG=chromium:523340
BRANCH=none
TEST="mosys eventlog list" no longer crashes on veyron_speedy

Change-Id: I16f507f4a4c65eeb28f062d5c041abdb08ffd838
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295207
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>