fdt: Modify API to return error codes instead of values
This makes the FDT library a bit more useful since we can return
an error code to indicate that reading the file failed (which likely
means we forgot to export a value from firmware) and sanity check
the value.

This will also make the API more uniform when we start reading other
types of data that can come from the devicetree.

BUG=none
BRANCH=none
TEST=ran "mosys memory spd print all" on veyron_mickey
Signed-off-by: David Hendricks <dhendrix@chromium.org>

Change-Id: I16d880604bab88a6051d1bbff62dc60195b3b7eb
Reviewed-on: https://chromium-review.googlesource.com/299765
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>