Create library for "generic" callbacks
This creates a location where we can put simple, generic callbacks
that can be used by platforms that do not need certain custom
callbacks to be created. The long-term goal is to reduce boilerplate
code.

To start, we'll target the psu_cb which for some platforms simply
returns an enum value to indicate if the platform uses a battery
or not.

BUG=none
BRANCH=none
TEST=tested on paine with follow-up patch

Change-Id: I72b90abe81d435b8eed1b61f53ca5372bc00eef4
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/315743
Reviewed-by: Shawn N <shawnn@chromium.org>