flashrom: Handle child process exit status properly
The "status" argument of waitpid() is assumed to be zeroed out if the
child process succeeds. However, this behavior isn't guaranteed, or at
least I didn't see any mention in the man page.

This patch changes this to first check if waitpid() returns >0 to
indicate that the process changed state and that waitpid() itself
didn't encounter any errors, and then checks if the exit code from
the child process was non-zero.

(credit to shawnn@ for pointing this out)

BUG=none
BRANCH=none
TEST=tested on cyan (gave it a bad flashrom binary)

Change-Id: I008a0900aecbf7ba6870230c10d141dd1c0e8526
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/318370
Reviewed-by: Shawn N <shawnn@chromium.org>