eeprom: Change return type of "size" function to int
This is intended to allow error handling in case the "size" function
fails. Up to this point we only worked with hard-coded values so error
checking didn't matter. This will make it so that we can leverage
external programs (flashrom and possibly others) and handle errors
if need be.

BUG=none
BRANCH=none
TEST=tested on cyan w/ follow-up patch, compiled for ARM and x86
platforms

Change-Id: Ib9251261b466281d5b8cda74b296d8e4d60acc61
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/319649
Reviewed-by: Shawn N <shawnn@chromium.org>