Add file lock mechanism
Some systems, such as Android, do not support SysV semaphore locks.
This implements an alternative file lock mechanism using flock().

flock() was chosen because it's pretty straight forward. It's known to
be broken when using NFS, but I doubt we'll ever store our lock on an
NFS volume.

For now the SysV semaphore lock is still in the codebase. It's
arguably better than the file lock and mosys was designed to be
flexible beyond the needs of ChromeOS, though in reality the SysV
semaphore lock probably won't be used anymore.

CQ-DEPEND=CL:329430,CL:327407
BUG=chrome-os-partner:49527
BRANCH=none
TEST=tested on Smaug by running mosys while flashrom was reading
firmware ROM, mosys paused until flashrom finished

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I82a7ac648e011ccbf1a872495dedb41217946c4a
Reviewed-on: https://chromium-review.googlesource.com/325609