nonspd: Remove incorrect entry for Hynix H9CCNNN8GTMLAR
There are two structs with different names for this module. The one
with the incorrect name has the correct module density specified.

This patch gets rid of the incorrect entry. The next patch will get
rid of the size embedded in the name, which will have the side-effect
of making the correct entry have a sensible name as well.

The incorrect entry is no longer used and can be safely removed.

BUG=none
BRANCH=none
TEST=compiled for oak and cyan

Change-Id: I0e85c2a7da3ecd7f8bac2b326c2408bac838377d
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/333006
Reviewed-by: Shawn N <shawnn@chromium.org>