nonspd: Remove size from name of nonspd module structs
As we added more modules, the size information embedded in the struct
names and the actual value for the size (.module_size_mbits) became
inconsistent. It appears that in many of these cases the total RAM
in a given SKU of a machine was put into the name while the module
density is used in the definition.

This simplifies the naming to remove the redundant and misleading
size info from the name.

BUG=none
BRANCH=none
TEST=compiled for ARM and x86 platforms

Change-Id: I6ad3c99fd9d39553bcd2cf15460184fabf83f8f2
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/333007
Reviewed-by: Shawn N <shawnn@chromium.org>