cros_ec: Point to separate address structs
This patch splits up the CrOS EC private data address members so they
are pointers to independent structs rather than values in a union.
This will allow generic code in upcoming patches to easily test if an
interface should be used based on its address struct pointer being set
or not.

For the most part we're just using the generic CrOS EC devfs
interface so this doesn't add much complexity for other platforms.

BUG=none
BRANCH=none
TEST=tested "mosys ec info" on veyron_mighty, paine, daisy_snow

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I093caf63a6ad05dd6c17b750208c0e0ba78811d9
Reviewed-on: https://chromium-review.googlesource.com/431518
Reviewed-by: Shawn N <shawnn@chromium.org>